// packages/ioc/src/metadata/meta.d.ts
export interface ProvidersMetadata {
    /**
     * provider services of the class.
     *
     * @type {KeyValue<Token, Token>}
     */
    providers?: Provider[];
}
/**
 * property metadata
 *
 * @export
 * @interface PropMetadata
 */
export interface PropertyMetadata<T = any> extends Omit<Parameter<T>, 'name'> {
    /**
     * property type
     *
     * @type {SymbolType}
     */
    type?: AbstractType<T>;
    /**
     * property name
     *
     * @type {string}
     */
    propertyKey: string;
}
// ...
export interface ParameterMetadata<T = any> extends Parameter<T> {
}
// ...
export interface ProvidedInMetadata {
    /**
     * the token provided in.
     */
    providedIn?: AbstractType | 'root' | 'platform';
}
/**
 * class pattern metadata.
 */
export interface PatternMetadata {
    /**
     * static provider or not.
     */
    static?: boolean;
    /**
     * is singleton or not.
     *
     * @type {boolean}
     */
    singleton?: boolean;
    /**
     * class cache timeout when not used.
     *
     * @type {number}
     */
    expires?: number;
}
/***
 * Annotation metadata.
 */
export interface AnnotationMetadata extends ProvidedInMetadata, ProvidersMetadata, PatternMetadata {
    /**
     * is abstract or not.
     */
    abstract?: boolean;
    /**
     * this type provider to.
     *
     * @type {SymbolType}
     */
    provide?: Token;
}